// src/app/shared/enums/status-processo.enum.ts
export enum StatusProcessoEnum {
    PROCESSO_CRIADO = 1,
    RECEBENDO_DOCUMENTACAO_TITULAR = 2,
    SOB_ANALISE_MEDICO = 3,
    SOB_ANALISE_DENTISTA = 4,
    SOB_ANALISE_ENFERMEIRO = 5,
    SOB_ANALISE_PSICOLOGO = 6,
    SOB_ANALISE_FONOAUDIOLOGO = 7,
    SOB_ANALISE_FISIOTERAPEUTA = 8,
    SOB_ANALISE_EQUIPE_TEC_ADMINISTRATIVA = 9,
    SOB_ANALISE_PROFISSIONAL_CAIXA = 10,
    AGUARD_DOC_COMPLEMENTAR = 11,
    PEDIDO_DEFERIDO = 12,
    PEDIDO_DEFERIDO_AUTOMATICAMENTE = 13,
    PEDIDO_INDEFERIDO = 14,
    INDEFERIMENTO_RATIFICADO = 15,
    CANCELADO_PELO_TITULAR = 16,
    PROCESSO_ENCERRADO = 17,
    REVISAO_REQUERIDA_PELO_TITULAR = 18,
    AGUARDANDO_COTACAO_OPME = 19,
    AUTORIZ_PREVIA_NEGADA = 20,
    AUTORIZACAO_PREVIA = 21,
    AUTORIZACAO_PREVIA_LIB_PARC = 22,
    AGUARD_CONS_PROFISSIONAL = 23,
    AGUARD_MANIF_PROFISSIONAL_ASSIT = 24,
    AGUARD_MANIF_PROFISSIONAL_DESMP = 25,
    BENEF_CONV_PERICIA = 26,
    AGUARD_PERICIA_PRESENCIAL = 27,
    AGUARD_RESULTADO_EXAMES = 28,
    AGUARD_PARECER_DESEMPATADOR = 29,
    AGUARD_DECISAO_ADM_FILIAL = 30,
    AGUARD_DECISAO_ADM_MATRIZ = 31,
    SOLICITADO_PARECER_COMPLEMENTAR_AUDITORIA = 32,
    SOLIC_FINALIZ_OP_AUDITORIA_LIBERACAO = 33,
    SOLIC_FINALIZ_OP_AUDITORIA_LIBERACAO_PARC = 34,
    SOLIC_FINALIZ_OP_AUDITORIA_NEGACAO = 35,
    SOB_AUDITORIA_INTERNA = 36,
    MIGRAR_SIST_SAUDE = 37,
    EM_PROCESSAMENTO_SIST_SAUDE = 38,
    FALHA_PROC_SIST_SAUDE = 39,
    EM_CONFERENCIA_SIST_SAUDE = 40,
    AGUARD_CAD_PROFISSIONAL_LIVRE = 41,
    CADASTRADO_SISTEMA_SAUDE = 42,
    CADASTRADO_SISTEMA_SAUDE_CARTAO_NAO_EMITIDO = 43,
    CANCELADO_NO_SISTEMA_DE_SAÚDE = 44,
 	ANALISE_CONCLUIDA_REEMBOLSO_INTEGRAL = 45,
    ANALISE_CONCLUIDA_GARANTIA_DE_ATENDIMENTO = 46,
    AGUARDANDO_LIBERACAO_DE_ALCADA_SUPERVISOR_CESAD = 47,
	AGUARDANDO_LIBERACAO_DE_ALCADA_COORDENADOR_CESAD = 48,
	AGUARDANDO_LIBERACAO_DE_ALCADA_GERENTE_DE_FILIAL_CESAD = 49,
	AGUARDANDO_LIBERACAO_DE_ALCADA_GERENTE_NACIONAL_GESAD = 50,
	AGUARDANDO_LIBERACAO_DE_ALCADA_SUPERINT_NACIONAL_SUBEN = 51,
    ALCADA_LIBERADA = 52,
    ALCADA_NAO_LIBERADA = 53
}